Separate responsibilities between Renderer and DefaultRendereringStrategy
Updated the Renderer to act in the same manner as the logic from the
DefaultRendereringStrategy. Once accomplished, updated the
DefaultRendereringStrategy to consume a Renderer in order to render a
console view model. Added a factory for the DefaultRendereringStrategy
to ensure it gets a Renderer, and added service configuration for the
renderer.

Since the console renderer does not render PHP view scripts, the limitations on
member variables are no longer required, either, and several methods (such as
`init()`) could be removed.