Fix test requirements
zend-mvc no longer requires zend-form, but does still list it as a development
dependency. It's used in the service configuration, however, which causes some
tests in this component to break. Updating to add a dev requirement on zend-form
solves the testing issues here.

I've also added a TODO file to track things to change before a stable release.